BE_py/book-platform/recommendation-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
mysql-connector-python
pydantic
scikit-learn
pandas
numpy
qdrant-client